+++ Functions/Public/Get-AppConnectionStrings.ps1
@@ -12,7 +12,7 @@
 .EXAMPLE
     PS> $DBConnectionString = Get-AppConnectionStrings
 .NOTES
-    0.9 (testare dopo refactoring)
+    1.0 (testato dopo refactoring)
     TODO: Criptare la password, qui appare in plain text
 #>